+++ src/main/java/com/structurizr/dsl/IncludeParser.java
@@ -4,27 +4,28 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
 final class IncludeParser extends AbstractParser {
 
-    private static final String GRAMMAR = "!include <file|url>";
+    private static final String GRAMMAR = "!include <file|directory|url>";
 
-    private static final int FILE_INDEX = 1;
+    private static final int SOURCE_INDEX = 1;
 
     void parse(IncludedDslContext context, Tokens tokens) {
-        // !include <file|url>
+        // !include <file|directory|url>
 
-        if (tokens.hasMoreThan(FILE_INDEX)) {
+        if (tokens.hasMoreThan(SOURCE_INDEX)) {
             throw new RuntimeException("Too many tokens, expected: " + GRAMMAR);
         }
 
-        if (!tokens.includes(FILE_INDEX)) {
+        if (!tokens.includes(SOURCE_INDEX)) {
             throw new RuntimeException("Expected: " + GRAMMAR);
         }
 
-        String source = tokens.get(FILE_INDEX);
+        String source = tokens.get(SOURCE_INDEX);
         if (source.startsWith("https://")) {
             String dsl = readFromUrl(source);
             List<String> lines = Arrays.asList(dsl.split("\n"));
@@ -32,24 +33,39 @@ final class IncludeParser extends AbstractParser {
             context.setFile(context.getFile());
         } else {
             if (context.getParentFile() != null) {
-                File file = new File(context.getParentFile().getParent(), source);
+                File path = new File(context.getParentFile().getParent(), source);
 
                 try {
-                    if (!file.exists()) {
-                        throw new RuntimeException(file.getCanonicalPath() + " could not be found");
+                    if (!path.exists()) {
+                        throw new RuntimeException(path.getCanonicalPath() + " could not be found");
                     }
 
-                    if (file.isDirectory()) {
-                        throw new RuntimeException(file.getCanonicalPath() + " should be a single file");
-                    }
-
-                    List<String> lines = Files.readAllLines(file.toPath(), StandardCharsets.UTF_8);
+                    List<String> lines = readLines(path);
                     context.setLines(lines);
-                    context.setFile(file);
+                    context.setFile(path);
                 } catch (IOException e) {
                     throw new RuntimeException(e.getMessage());
                 }
             }
+        }
+    }
+
+    private List<String> readLines(File path) throws IOException {
+        if (path.isDirectory()) {
+            List<String> lines = new ArrayList<>();
+
+            File[] files = path.listFiles();
+            if (files != null) {
+                Arrays.sort(files);
+
+                for (File file : files) {
+                    lines.addAll(readLines(file));
+                }
+            }
+
+            return lines;
+        } else {
+            return Files.readAllLines(path.toPath(), StandardCharsets.UTF_8);
         }
     }